pyproject.toml
[project]
name = "travelagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "litellm>=1.73.1",
    "openai-agents>=0.0.19",
    "py-dotenv>=0.1",
]
